Added the x64 platform to the development solution. Changed the dll-keyword-handling. Added PASTEL_DYNAMIC_LIBRARIES and PASTEL_LARGE_INTEGER. Changed the definition of 'integer'. Updated the build system. Updated SecondIterator and SparseIterator. Updated Rational class to work with different kinds of native integers.

+++ pastel/geometry/search_all_neighbors_pointkdtree.hpp
@@ -100,7 +100,7 @@ namespace Pastel
 						nearestSet.begin() + nearestCount,
 						nearestArray->rowBegin(i));
 				}
-				const integer fillStart = std::max(nearestCount - kNearestBegin, 0);
+				const integer fillStart = std::max(nearestCount - kNearestBegin, (integer)0);
 				std::fill(
 					nearestArray->rowBegin(i) + fillStart,
 					nearestArray->rowEnd(i),
@@ -115,7 +115,7 @@ namespace Pastel
 						distanceSet.begin() + nearestCount,
 						distanceArray->rowBegin(i));
 				}
-				const integer fillStart = std::max(nearestCount - kNearestBegin, 0);
+				const integer fillStart = std::max(nearestCount - kNearestBegin, (integer)0);
 				std::fill(
 					distanceArray->rowBegin(i) + fillStart,
 					distanceArray->rowEnd(i), infinity<Real>());
